Moved modeladmin check tests to a separate file.
